requirements.txt
mcp
notion-client
python-dotenv